Finished locked door tutorial
This is a long commit but that's because I spent so long trying to rework the tutorial code into my own. I had a different way of opening and closing a door that I wanted to keep. So I spent a lot of time on that instead of several commits.