Allow players to take inventory objects, modify Door script to take keys into account,
This is a lengthy commit, again, thanks to my frustration. I constantly need to rework the code in the tutorials to fit my doors. At the moment David, my teacher, has his script so that the door can only be opened and never closed. I personally don't like that, so I made it possible to be closed again. The problem is that it works. Just not entirely. The door's display text is stuck on "Locked Door" I'll have to fix that soon. Otherwise the door works with keys.